Added exportAssembly function
The exportAssembly function combines all visible objects into a single
compound object to a new document. This function is intended to
simplify the task of use a PCB assembly with FreeCAD Assembly workbench.
This function also implements a trick to deal with a FreeCAD bug that
causes trouble making compound with holes in overlapped objects.